requirements.txt
flask==2.3.2
flask-cors==3.0.10
torch==2.7.0
torchvision==0.22.0
pillow==10.0.0
